+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### documentdb v0.111-0 (Unreleased) ###
+* Add `--enable-http` flag for `documentdb-local` to disable TLS enforcement on the gateway endpoint *[Feature]*
 
 ### documentdb v0.110-0 (Unreleased) ###
 * Add support for keyword `description` in `$jsonSchema` *[Feature]*